requirements.txt
google-generativeai>=0.8.0
gspread>=6.0.0
gspread-formatting>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
markdown>=3.5.0
pytest>=8.0.0
pytest-mock>=3.12.0
